requirements.txt
flask==2.2.5
opencv-python==4.5.5.64
Pillow==9.5.0
